microservices/library-service/requirements.txt
fastapi[all]
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
aioredis
prometheus-fastapi-instrumentator
python-json-logger
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-otlp
alembic
boto3
aioboto3
python-multipart
celery
redis
httpx
